requirements.txt
web3==6.11.3
pandas==2.1.3
numpy==1.25.2
python-dotenv==1.0.0
requests==2.31.0
eth-account==0.9.0
eth-abi==4.2.1
tqdm==4.66.1

# New ML dependencies for multi-agent framework
torch==2.1.0
statsmodels==0.14.0
scikit-learn==1.3.0
plotly==5.17.0
streamlit==1.28.0

# Optional but recommended
jupyter==1.0.0
matplotlib==3.8.0
seaborn==0.13.0 